+++ gulpfile.js
@@ -3,9 +3,15 @@ var $ = require('gulp-load-plugins')();
 var streamqueue = require('streamqueue');
 var open = require('open');
 var runSequence = require('run-sequence');
+var bowerFiles = require('main-bower-files');
+var series = require('stream-series');
+
+// TEACHSCAPE
+// var localReflect = ('../reflect/ts-angular-bootstrap-calendar/');
 
 gulp.task('watch', ['server'], function() {
   $.livereload.listen();
+  gulp.start('test:watch');
   gulp.watch('src/less/*.less', ['less']);
   gulp.watch('src/**/*.js', ['lint']);
   gulp.watch('css/*.css').on('change', $.livereload.changed);
@@ -14,7 +20,7 @@ gulp.task('watch', ['server'], function() {
     './docs/scripts/*.js',
     './docs/styles/*.css',
     './src/**/*.js',
-    './templates/**']
+    './src/templates/**']
   ).on('change', $.livereload.changed);
 });
 
@@ -35,15 +41,26 @@ gulp.task('less', function() {
     .pipe(gulp.dest('css'))
 });
 
+var pkg = require('./bower.json');
+var banner = ['/**',
+  ' * <%= pkg.name %> - <%= pkg.description %>',
+  ' * @version v<%= pkg.version %>',
+  ' * @link <%= pkg.homepage %>',
+  ' * @license <%= pkg.license %>',
+  ' */',
+  ''].join('\n');
+
 gulp.task('css', function() {
 
   return gulp.src('src/less/calendar.less')
     .pipe($.sourcemaps.init())
     .pipe($.less())
-    .pipe($.rename('angular-bootstrap-calendar.css'))
+    .pipe($.header(banner, { pkg : pkg } ))
+    .pipe($.rename('ts-angular-bootstrap-calendar.css'))
     .pipe(gulp.dest('dist/css'))
     .pipe($.minifyCss())
-    .pipe($.rename('angular-bootstrap-calendar.min.css'))
+    .pipe($.rename('ts-angular-bootstrap-calendar.min.css'))
+    .pipe($.header(banner, { pkg : pkg } ))
     .pipe($.sourcemaps.write('.'))
     .pipe(gulp.dest('dist/css'));
 
@@ -52,7 +69,7 @@ gulp.task('css', function() {
 function getTemplates() {
 
   return gulp
-    .src('templates/**/*.html')
+    .src('src/templates/**/*.html')
     .pipe($.htmlmin({
       removeComments: true,
       collapseWhitespace: true
@@ -60,7 +77,7 @@ function getTemplates() {
     .pipe($.angularTemplatecache({
       standalone: false,
       module: 'mwl.calendar',
-      root: 'templates/'
+      root: 'src/templates/'
     }));
 
 }
@@ -71,8 +88,8 @@ function mergeStreams(stream1, stream2) {
 
 function buildJS(withTemplates) {
 
-  var minFilename = withTemplates ? 'angular-bootstrap-calendar-tpls.min.js' : 'angular-bootstrap-calendar.min.js';
-  var unminfilename = withTemplates ? 'angular-bootstrap-calendar-tpls.js' : 'angular-bootstrap-calendar.js';
+  var minFilename = withTemplates ? 'ts-angular-bootstrap-calendar-tpls.min.js' : 'ts-angular-bootstrap-calendar.min.js';
+  var unminfilename = withTemplates ? 'ts-angular-bootstrap-calendar-tpls.js' : 'ts-angular-bootstrap-calendar.js';
 
   var stream = withTemplates ? mergeStreams(
     gulp.src('src/**/*.js'),
@@ -86,9 +103,11 @@ function buildJS(withTemplates) {
     .pipe($.ngAnnotate())
     .pipe($.concat(unminfilename))
     .pipe($.wrapJs('(function(window, angular) {\n%= body %\n }) (window, angular);'))
+    .pipe($.header(banner, { pkg : pkg } ))
     .pipe(gulp.dest('dist/js'))
     .pipe($.uglify())
     .pipe($.rename(minFilename))
+    .pipe($.header(banner, { pkg : pkg } ))
     .pipe($.sourcemaps.write('.'))
     .pipe(gulp.dest('dist/js'));
 }
@@ -102,6 +121,7 @@ gulp.task('js', ['js-tpls'], function() {
 });
 
 gulp.task('build', ['js', 'css'], function() {});
+gulp.task('dev', ['js', 'css'], function() {});
 
 function release(importance) {
   return gulp.src(['./package.json', './bower.json'])
@@ -109,9 +129,6 @@ function release(importance) {
     .pipe(gulp.dest('./'));
 }
 
-gulp.task('release:patch', function() { return release('patch'); })
-gulp.task('release:minor', function() { return release('minor'); })
-gulp.task('release:major', function() { return release('major'); })
 
 gulp.task('default', ['watch'], function() {});
 
@@ -135,6 +152,40 @@ gulp.task('ci:lint', function() {
   return lint(true);
 });
 
+function runTests(action, onDistCode) {
+  var vendorJs = gulp.src(bowerFiles({includeDev: true})).pipe($.filter('*.js'));
+  if (onDistCode) {
+    var appJs = gulp.src('dist/js/ts-angular-bootstrap-calendar-tpls.min.js');
+  } else {
+    var appJs = gulp.src('src/**/*.js').pipe($.sort()).pipe($.angularFilesort());
+  }
+  var templates = gulp.src('src/templates/*.html');
+  var karmaSetup = gulp.src('test/karma.setup.js');
+  var test = gulp.src('test/unit/**/*.js');
+
+  return series(vendorJs, appJs, templates, karmaSetup, test)
+    .pipe($.karma({
+      configFile: 'karma.conf.js',
+      action: action
+    }));
+}
+
+gulp.task('test:src', function() {
+  return runTests('run').on('error', function(err) {
+    throw err;
+  });
+});
+
+gulp.task('test:dist', function() {
+  return runTests('run', true).on('error', function(err) {
+    throw err;
+  });
+});
+
+gulp.task('test:watch', function() {
+  return runTests('watch');
+});
+
 gulp.task('ci', function(done) {
-  runSequence('ci:lint', 'build', done);
+  runSequence('ci:lint', 'build', 'test:dist', done);
 });
